feat: domain syntax highlighting
